+++ data_linter/validators/pandas_validator.py
@@ -2,6 +2,7 @@ import logging
 import inspect
 from functools import wraps
 from datetime import datetime
+from mojap_metadata import Metadata
 from typing import Union
 
 import pandas as pd
@@ -20,6 +21,8 @@ log = logging.getLogger("root")
 default_date_format = "%Y-%m-%d"
 default_datetime_format = "%Y-%m-%d %H:%M:%S"
 
+class ColumnError(Exception):
+    pass
 
 class PandasValidator(BaseTableValidator):
     """
@@ -59,26 +62,16 @@ class PandasValidator(BaseTableValidator):
         Data is read using pd_arrow_parser.
         """
 
-        df = _parse_data_to_pandas(self.filepath, self.table_params, self.metadata)
+        df, self.metadata = _parse_data_to_pandas(
+            self.filepath, self.table_params, self.metadata
+        )
         self.validate_df(df)
 
     def get_response_dict(self):
         return self.response.get_result()
 
     def validate_df(self, df):
-
-        meta_cols = [col for col in self.metadata["columns"] if col["name"] in df]
-
-        cols_not_tested = [
-            col for col in self.metadata["columns"] if col["name"] not in df
-        ]
-        if cols_not_tested:
-            log.info(
-                "some columns will not be tested as not present in metadata: "
-                f"{cols_not_tested}"
-            )
-
-        for m in meta_cols:
+        for m in self.metadata["columns"]:
             self.validate_col(df[m["name"]], m)
 
     def validate_col(self, col, meta_col):
@@ -398,7 +391,7 @@ def _parse_data_to_pandas(filepath: str, table_params: dict, metadata: dict):
     if "csv" in metadata["file_format"]:
         df = reader.read_csv(filepath, dtype=str, low_memory=False)
     elif "json" in metadata["file_format"]:
-        df = reader.read_json(filepath, lines=True)
+        df = reader.read_json(filepath, dtype=str, lines=True)
     elif "parquet" in metadata["file_format"]:
         df = pa_read_parquet_to_pandas(filepath)
         data_is_not_parquet = False
@@ -411,6 +404,50 @@ def _parse_data_to_pandas(filepath: str, table_params: dict, metadata: dict):
             c["name"] = c["name"].lower()
         df.columns = [c.lower() for c in df.columns]
 
+
+    allow_missing_cols = table_params.get("allow-missing-cols", False)
+    allow_unexpected_data = table_params.get("allow-unexpected-data", False)
+
+    meta_col_names = [ 
+        c["name"] for c in metadata["columns"]
+        if c["name"] not in metadata.get("partitions", [])
+    ]
+    cols_in_meta_but_not_data = [c for c in meta_col_names if c not in df.columns]
+    cols_in_data_but_not_meta = [c for c in df.columns if c not in meta_col_names]
+    cols_in_data_and_meta = [c for c in df.columns if c in meta_col_names]
+
+    # error if there are no common columns
+    if not cols_in_data_and_meta:
+        raise ColumnError("There is no commonality between the data and metadata")
+
+    raise_column_error = False
+    err_msg = ""
+
+    # remove columns from meta that aren't in the data if allowed
+    msg_1 = f"columns present in metadata but not in data: {cols_in_meta_but_not_data}"
+    if not allow_missing_cols and cols_in_meta_but_not_data:
+        err_msg += msg_1
+        raise_column_error = True
+    elif allow_missing_cols and cols_in_meta_but_not_data:
+        meta_tmp = Metadata.from_dict(metadata)
+        for col in cols_in_meta_but_not_data:
+            meta_tmp.remove_column(col)
+        metadata = meta_tmp.to_dict()
+        log.info("not testing " + msg_1)
+    
+    # error if there is unexepcted data, unless allowed
+    msg_2 = f"columns present in data but not in metadata: {cols_in_data_but_not_meta}"
+    if not allow_unexpected_data and cols_in_data_but_not_meta:
+        err_msg += f"\n{msg_2}"
+        raise_column_error = True
+    elif allow_unexpected_data and cols_in_data_but_not_meta:
+        log.info("not testing " + msg_2)
+        df = df[cols_in_data_and_meta]
+
+    # raise the error with all details, if required
+    if raise_column_error:
+        raise ColumnError(err_msg)
+
     # cast table column by column if it's not parquet, except timestamps
     if data_is_not_parquet:
         for c in metadata["columns"]:
@@ -420,16 +457,7 @@ def _parse_data_to_pandas(filepath: str, table_params: dict, metadata: dict):
     if table_params.get("row-limit"):
         df = df.sample(table_params.get("row-limit"))
 
-    if table_params.get("only-test-cols-in-metadata", False):
-        meta_col_names = [
-            c["name"]
-            for c in metadata["columns"]
-            if c["name"] not in metadata.get("partitions", [])
-        ]
-        keep_cols = [c for c in df.columns if c in meta_col_names]
-        df = df[keep_cols]
-
-    return df
+    return df, metadata
 
 
 def _check_pandas_series_is_str(s: pd.Series, na_as=True):